+++ cookbooks/aws-parallelcluster-entrypoints/recipes/update_computefleet_status.rb
@@ -12,5 +12,10 @@
 # OR CONDITIONS OF ANY KIND, express or implied. See the License for the specific language governing permissions and
 # limitations under the License.
 
+chef_handler 'ErrorHandlers::UpdateFailureHandler' do
+  arguments start_clustermgtd: true
+  type exception: true
+end
+
 include_recipe 'aws-parallelcluster-shared::setup_envars'
 include_recipe 'aws-parallelcluster-computefleet::update_computefleet_status'