+++ CHANGELOG.md
@@ -31,7 +31,7 @@ This file is used to list changes made in each version of the AWS ParallelCluste
 - Upgrade mysql-community-client to version 8.4.8 (from 8.0.39) for all OSs except Amazon Linux 2.
 - Upgrade Intel MPI Library to 2021.17.2 (from 2021.16.0).
 - Upgrade Cinc Client to version 18.8.54 (from 18.7.10).
-- Always start clustermgtd on cluster update failure, regardless the failure condition.
+- Always start clustermgtd on cluster update and compute fleet status update failure, regardless the failure condition.
 
 **BUG FIXES**
 - Fix timestamp formats in CloudWatch log configuration to let CloudWatch parse the correct timestamps.